separated into different files, added comments
